Rename spotdl.command_line.lib to spotdl.command_line.core
