+++ spotdl/__init__.py
@@ -1,5 +1,5 @@
 from spotdl.command_line.__main__ import main
 
 from spotdl.version import __version__
-from spotdl.command_line.lib import Spotdl
+from spotdl.command_line.core import Spotdl
